+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/State.hs
@@ -4,4 +4,5 @@ module Cardano.Ledger.Alonzo.State (
 
 import Cardano.Ledger.Mary.State
 
+import Cardano.Ledger.Alonzo.State.CertState ()
 import Cardano.Ledger.Alonzo.State.Stake ()