+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Binary/RoundTrip.hs
@@ -12,7 +12,6 @@ module Test.Cardano.Ledger.Shelley.Binary.RoundTrip (
 ) where
 
 import Cardano.Ledger.Binary
-import Cardano.Ledger.CertState
 import Cardano.Ledger.Compactible
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley (ShelleyEra)