+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Cert.hs
@@ -15,15 +15,13 @@ module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert () where
 
 import Cardano.Ledger.Address (RewardAccount)
 import Cardano.Ledger.BaseTypes
-import Cardano.Ledger.CertState
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.Rules
+import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Conway.TxCert
-import Cardano.Ledger.Shelley.CertState
 import Cardano.Ledger.Shelley.LedgerState
-import Cardano.Ledger.State
 import qualified Data.Foldable as Set
 import Data.Functor.Identity (Identity)
 import Data.Map.Strict (Map, keysSet)
@@ -58,13 +56,13 @@ instance
       <*> toSpecRep withdrawals
       <*> toSpecRep ccColdCreds
 
-instance SpecTranslate ctx (ShelleyCertState era) where
-  type SpecRep (ShelleyCertState era) = Agda.CertState
-  toSpecRep ShelleyCertState {..} =
+instance SpecTranslate ctx (ConwayCertState era) where
+  type SpecRep (ConwayCertState era) = Agda.CertState
+  toSpecRep ConwayCertState {..} =
     Agda.MkCertState
-      <$> toSpecRep shelleyCertDState
-      <*> toSpecRep shelleyCertPState
-      <*> toSpecRep shelleyCertVState
+      <$> toSpecRep conwayCertDState
+      <*> toSpecRep conwayCertPState
+      <*> toSpecRep conwayCertVState
 
 instance Era era => SpecTranslate ctx (ConwayTxCert era) where
   type SpecRep (ConwayTxCert era) = Agda.DCert
@@ -78,7 +76,7 @@ instance
   , SpecRep (TxOut era) ~ Agda.TxOut
   , GovState era ~ ConwayGovState era
   , Inject ctx (CertState era)
-  , EraCertState era
+  , ConwayEraCertState era
   ) =>
   SpecTranslate ctx (UTxOState era)
   where
@@ -104,8 +102,8 @@ instance
   , GovState era ~ ConwayGovState era
   , SpecTranslate (CertState era) (TxOut era)
   , SpecRep (CertState era) ~ Agda.CertState
-  , EraCertState era
-  , CertState era ~ ShelleyCertState era
+  , ConwayEraCertState era
+  , CertState era ~ ConwayCertState era
   ) =>
   SpecTranslate ctx (LedgerState era)
   where
@@ -133,8 +131,8 @@ instance
   , GovState era ~ ConwayGovState era
   , SpecTranslate (CertState era) (TxOut era)
   , SpecRep (CertState era) ~ Agda.CertState
-  , EraCertState era
-  , CertState era ~ ShelleyCertState era
+  , ConwayEraCertState era
+  , CertState era ~ ConwayCertState era
   ) =>
   SpecTranslate ctx (EpochState era)
   where
@@ -213,8 +211,8 @@ instance
   , GovState era ~ ConwayGovState era
   , SpecTranslate (CertState era) (TxOut era)
   , SpecRep (CertState era) ~ Agda.CertState
-  , EraCertState era
-  , CertState era ~ ShelleyCertState era
+  , ConwayEraCertState era
+  , CertState era ~ ConwayCertState era
   ) =>
   SpecTranslate ctx (NewEpochState era)
   where